Update _flex-container.scss
Added $parent and $children selectors. I've been thinking about this for a while, since we are technically instantiating an object that contains children, it would be nice to have the option of configuring both the container and the children. This lets us be more specific if needed. (I'm also open to naming `$parent` as `$container` instead, if you feel it would make more sense)

The default & { $things } basically applies to the current CSS item.