298-change "sticky-footer" mixin to target <footer> selector  (#299)
remove parent <footer> element from the Global Footer component and include it instead in kss/builder/decanter/index.twig.  The Global Footer component is now a <section> to be included in a a sites <footer> region.  Also added some comments to global-footer.twig which mirror what are in the brand-bar component.